+++ docs/source/changes.rst
@@ -17,6 +17,8 @@ all releases are available on `PyPI <https://pypi.org/project/pytask>`_ and
 - :gh:`150` adds a limit on the number of items displayed in the execution table which
   is also configurable with ``--n-entries-in-table`` on the cli and
   ``n_entries_in_table`` in the configuration file.
+- :gh:`152` makes the duration of the execution readable by humans by separating it into
+  days, hours, minutes and seconds.
 
 
 0.1.1 - 2021-08-25